API Reference

HypothesisTests.OneSampleTTest — Type

OneSampleTTest(r::ComposedEstimate, Ψ₀=0)

Performs a T test on the ComposedEstimate.

HypothesisTests.OneSampleTTest — Type

OneSampleTTest(r::AsymptoticallyLinearEstimate, Ψ₀=0)

Performs a T test on the AsymptoticallyLinearEstimate.

HypothesisTests.OneSampleZTest — Type

OneSampleZTest(r::ComposedEstimate, Ψ₀=0)

Performs a T test on the ComposedEstimate.

HypothesisTests.OneSampleZTest — Type

OneSampleZTest(r::AsymptoticallyLinearEstimate, Ψ₀=0)

Performs a Z test on the AsymptoticallyLinearEstimate.

TMLE.AverageTreatmentEffect — Type

Average Treatment Effect / ATE

Definition

$ATE(Y, T, case, control) = E[Y|do(T=case)] - E[Y|do(T=control)$

Constructors

ATE(scm::SCM; outcome::Symbol, treatment::NamedTuple)

ATE( outcome::Symbol, treatment::NamedTuple, confounders::Union{Symbol, AbstractVector{Symbol}}, covariates::Union{Nothing, Symbol, AbstractVector{Symbol}} = nothing, outcomemodel = withencoder(LinearRegressor()), treatment_model = LinearBinaryClassifier()

)

Example

Ψ = ATE(scm, outcome=:Y, treatment=(T=(case=1,control=0),)

TMLE.BackdoorAdjustment — Method

BackdoorAdjustment(;outcome_extra=[])

The adjustment set for each treatment variable is simply the set of direct parents in the associated structural model.

outcome_extra are optional additional variables that can be used to fit the outcome model in order to improve inference.

TMLE.ConditionalMean — Type

Conditional Mean / CM

Definition

$CM(Y, T=t) = E[Y|do(T=t)]$

Constructors

CM(scm::SCM; outcome::Symbol, treatment::NamedTuple)

CM( outcome::Symbol, treatment::NamedTuple, confounders::Union{Symbol, AbstractVector{Symbol}}, covariates::Union{Nothing, Symbol, AbstractVector{Symbol}} = nothing, outcomemodel = withencoder(LinearRegressor()), treatment_model = LinearBinaryClassifier()

)

Example

Ψ = CM(scm, outcome=:Y, treatment=(T=1,))

TMLE.InteractionAverageTreatmentEffect — Type

Interaction Average Treatment Effect / IATE

Definition

For two treatments with case/control settings (1, 0):

$IATE = E[Y|do(T₁=1, T₂=1)] - E[Y|do(T₁=1, T₂=0)] - E[Y|do(T₁=0, T₂=1)] + E[Y|do(T₁=0, T₂=0)]$

Constructors

IATE(scm::SCM; outcome::Symbol, treatment::NamedTuple)

IATE( outcome::Symbol, treatment::NamedTuple, confounders::Union{Symbol, AbstractVector{Symbol}}, covariates::Union{Nothing, Symbol, AbstractVector{Symbol}} = nothing, outcomemodel = withencoder(LinearRegressor()), treatment_model = LinearBinaryClassifier()

)

Example

Ψ = IATE(scm, outcome=:Y, treatment=(T₁=(case=1,control=0), T₂=(case=1,control=0))

TMLE.StructuralCausalModel — Type

Structural Causal Model / SCM

Constructors

SCM(;equations=Dict{Symbol, SE}()) SCM(equations::Vararg{SE})

Examples

scm = SCM( SE(:Y, [:T, :W, :C]), SE(:T, [:W]) )

TMLE.StructuralEquation — Type

Structural Equation / SE

Constructors

SE(outcome, parents; model=nothing)

SE(;outcome, parents, model=nothing)

Examples

eq = SE(:Y, [:T, :W]) eq = SE(:Y, [:T, :W], model = LinearRegressor())

TMLE.TreatmentTransformer — Method

TreatmentTransformer(;encoder=encoder())

Treatments in TMLE are represented by CategoricalArrays. If a treatment column has type OrderedFactor, then its integer representation is used, make sure that the levels correspond to your expectations. All other columns are one-hot encoded.

Distributions.estimate — Method

Distributions.estimate(r::AsymptoticallyLinearEstimate)

Retrieves the final estimate: after the TMLE step.

Distributions.estimate — Method

Distributions.estimate(r::ComposedEstimate)

Retrieves the final estimate: after the TMLE step.

Statistics.var — Method

var(r::AsymptoticallyLinearEstimate)

Computes the estimated variance associated with the estimate.

Statistics.var — Method

var(r::ComposedEstimate)

Computes the estimated variance associated with the estimate.

StatsAPI.fit! — Method

MLJBase.fit!(eq::SE, dataset; input_variables=nothing, verbosity=1, cache=true, force=false)

Fits the outcome's Structural Equation using the dataset with inputs variables given by either:

The variables corresponding to parents(eq) if input_variables= nothing

The alternative variables provided by input_variables.

Extra keyword arguments are:

cache: Controls whether the associated MLJ.Machine will cache data.

force: Controls whether to force the associated MLJ.Machine to refit even if neither the model or data has changed.

verbosity: Controls the verbosity level

TMLE.StaticConfoundedModel — Method

StaticConfoundedModel(
    outcome::Symbol, treatment::Symbol, confounders::Union{Symbol, AbstractVector{Symbol}}; 
    covariates::Union{Nothing, Symbol, AbstractVector{Symbol}} = nothing, 
    outcome_model = TreatmentTransformer() |> LinearRegressor(),
    treatment_model = LinearBinaryClassifier()
)

Defines a classic Structural Causal Model with one outcome, one treatment, a set of confounding variables and optional covariates influencing the outcome only.

The outcome_model and treatment_model define the relationship between the outcome (resp. treatment) and their ancestors.

TMLE.StaticConfoundedModel — Method

StaticConfoundedModel(
    outcomes::Vector{Symbol}, 
    treatments::Vector{Symbol}, 
    confounders::Union{Symbol, AbstractVector{Symbol}}; 
    covariates::Union{Nothing, Symbol, AbstractVector{Symbol}} = nothing, 
    outcome_model = TreatmentTransformer() |> LinearRegressor(),
    treatment_model = LinearBinaryClassifier()
)

Defines a classic Structural Causal Model with multiple outcomes, multiple treatments, a set of confounding variables and optional covariates influencing the outcomes only.

All treatments are assumed to be direct parents of all outcomes. The confounding variables are shared for all treatments.

The outcome_model and treatment_model define the relationships between the outcomes (resp. treatments) and their ancestors.

TMLE.compose — Method

compose(f, estimation_results::Vararg{AsymptoticallyLinearEstimate, N}) where N

Provides an estimator of f(estimation_results...).

Mathematical details

The following is a summary from Asymptotic Statistics, A. W. van der Vaart.

Consider k TMLEs computed from a dataset of size n and embodied by Tₙ = (T₁,ₙ, ..., Tₖ,ₙ). Since each of them is asymptotically normal, the multivariate CLT provides the joint distribution:

√n(Tₙ - Ψ₀) ↝ N(0, Σ),

where Σ is the covariance matrix of the TMLEs influence curves.

Let f:ℜᵏ→ℜᵐ, be a differentiable map at Ψ₀. Then, the delta method provides the limiting distribution of √n(f(Tₙ) - f(Ψ₀)). Because Tₙ is normal, the result is:

√n(f(Tₙ) - f(Ψ₀)) ↝ N(0, ∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ),

where ∇f(Ψ₀):ℜᵏ→ℜᵐ is a linear map such that by abusing notations and identifying the function with the multiplication matrix: ∇f(Ψ₀):h ↦ ∇f(Ψ₀) ̇h. And the matrix ∇f(Ψ₀) is the jacobian of f at Ψ₀.

Hence, the only thing we need to do is:

Compute the covariance matrix Σ

Compute the jacobian ∇f, which can be done using Julia's automatic differentiation facilities.

The final estimator is normal with mean f₀=f(Ψ₀) and variance σ₀=∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ

Arguments

f: An array-input differentiable map.

estimation_results: 1 or more AsymptoticallyLinearEstimate structs.

Examples

Assuming res₁ and res₂ are TMLEs:

f(x, y) = [x^2 - y, y - 3x]
compose(f, res₁, res₂)

TMLE.optimize_ordering! — Method

optimize_ordering!(estimands::Vector{<:Estimand})

Optimizes the order of the estimands to maximize reuse of fitted equations in the associated SCM.

TMLE.optimize_ordering — Method

optimize_ordering(estimands::Vector{<:Estimand})

See optimize_ordering!

TMLE.tmle! — Method

tmle!(Ψ::CMCompositeEstimand, dataset; 
    adjustment_method=BackdoorAdjustment(), 
    verbosity=1, 
    force=false, 
    ps_lowerbound=1e-8, 
    weighted_fluctuation=false
    )

Performs Targeted Minimum Loss Based Estimation of the target estimand.

Arguments

Ψ: An estimand of interest.

dataset: A table respecting the Tables.jl interface.

adjustment_method: A confounding adjustment method.

verbosity: Level of logging.

force: To force refit of machines in the SCM .

ps_lowerbound: The propensity score will be truncated to respect this lower bound.

weighted_fluctuation: To use a weighted fluctuation instead of the vanilla TMLE, can improve stability.