API Reference

HypothesisTests.OneSampleTTest — Type

OneSampleTTest(r::EICEstimate, Ψ₀=0)

Performs a T test on the EICEstimate.

HypothesisTests.OneSampleTTest — Type

OneSampleTTest(r::ComposedEstimate, Ψ₀=0)

Performs a T test on the ComposedEstimate.

HypothesisTests.OneSampleZTest — Type

OneSampleZTest(r::EICEstimate, Ψ₀=0)

Performs a Z test on the EICEstimate.

HypothesisTests.OneSampleZTest — Type

OneSampleZTest(r::ComposedEstimate, Ψ₀=0)

Performs a T test on the ComposedEstimate.

TMLE.SCM — Type

A SCM is simply a MetaGraph over a Directed Acyclic Graph with additional methods.

TMLE.TreatmentTransformer — Method

TreatmentTransformer(;encoder=encoder())

Treatments in TMLE are represented by CategoricalArrays. If a treatment column has type OrderedFactor, then its integer representation is used, make sure that the levels correspond to your expectations. All other columns are one-hot encoded.

Distributions.estimate — Method

Distributions.estimate(r::ComposedEstimate)

Retrieves the final estimate: after the TMLE step.

Distributions.estimate — Method

Distributions.estimate(r::EICEstimate)

Retrieves the final estimate: after the TMLE step.

Statistics.var — Method

var(r::ComposedEstimate)

Computes the estimated variance associated with the estimate.

Statistics.var — Method

var(r::EICEstimate)

Computes the estimated variance associated with the estimate.

TMLE.StaticSCM — Method

A plate Structural Causal Model where:

For all outcomes: oᵢ = fᵢ(treatments, confounders, outcomeextracovariates)

For all treatments: tⱼ = fⱼ(confounders)

Example

StaticSCM([:Y], [:T₁, :T₂], [:W₁, :W₂, :W₃]; outcomeextracovariates=[:C])

TMLE.compose — Method

compose(f, estimation_results::Vararg{EICEstimate, N}) where N

Provides an estimator of f(estimation_results...).

Mathematical details

The following is a summary from Asymptotic Statistics, A. W. van der Vaart.

Consider k TMLEs computed from a dataset of size n and embodied by Tₙ = (T₁,ₙ, ..., Tₖ,ₙ). Since each of them is asymptotically normal, the multivariate CLT provides the joint distribution:

√n(Tₙ - Ψ₀) ↝ N(0, Σ),

where Σ is the covariance matrix of the TMLEs influence curves.

Let f:ℜᵏ→ℜᵐ, be a differentiable map at Ψ₀. Then, the delta method provides the limiting distribution of √n(f(Tₙ) - f(Ψ₀)). Because Tₙ is normal, the result is:

√n(f(Tₙ) - f(Ψ₀)) ↝ N(0, ∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ),

where ∇f(Ψ₀):ℜᵏ→ℜᵐ is a linear map such that by abusing notations and identifying the function with the multiplication matrix: ∇f(Ψ₀):h ↦ ∇f(Ψ₀) ̇h. And the matrix ∇f(Ψ₀) is the jacobian of f at Ψ₀.

Hence, the only thing we need to do is:

Compute the covariance matrix Σ

Compute the jacobian ∇f, which can be done using Julia's automatic differentiation facilities.

The final estimator is normal with mean f₀=f(Ψ₀) and variance σ₀=∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ

Arguments

f: An array-input differentiable map.

estimation_results: 1 or more EICEstimate structs.

Examples

Assuming res₁ and res₂ are TMLEs:

f(x, y) = [x^2 - y, y - 3x]
compose(f, res₁, res₂)