API Reference

HypothesisTests.OneSampleTTest — Type

OneSampleTTest(r::ComposedTMLE, Ψ₀=0)

Performs a T test on the ComposedTMLE.

HypothesisTests.OneSampleTTest — Type

OneSampleTTest(r::ALEstimate, Ψ₀=0)

Performs a T test on the ALEstimate.

HypothesisTests.OneSampleZTest — Type

OneSampleZTest(r::ALEstimate, Ψ₀=0)

Performs a Z test on the ALEstimate.

TMLE.ATE — Type

ATE: Average Treatment Effect

Mathematical definition:

Eₓ[E[Y|do(T=case), X]] - Eₓ[E[Y|do(T=control), X]]

Assumed Causal graph:

 T  ←  W  
  ↘   ↙ 
    Y  ← C

Notation:

Y: target

T: treatment

W: confounders

C: covariates

X = (W, C, T)

Fields:

- target: A symbol identifying the target variable of interest
- treatment: A NamedTuple linking each treatment variable to case/control values
- confounders: Confounding variables affecting both the target and the treatment
- covariates: Optional extra variables affecting the target only

Examples:

ATE₁ = ATE(
    target=:Y₁,
    treatment=(T₁=(case=1, control=0),),
    confounders=[:W₁, :W₂],
    covariates=[:C₁]
)

ATE₂ = ATE(
    target=:Y₂,
    treatment=(T₁=(case=1, control=0), T₂=(case="A", control="B")),
    confounders=[:W₁],
)

TMLE.CM — Type

CM: Conditional Mean

Mathematical definition:

Eₓ[E[Y|do(T=t), X]]

Assumed Causal graph:

 T  ←  W  
  ↘   ↙ 
    Y  ← C

Notation:

Y: target

T: treatment

W: confounders

C: covariates

X = (W, C, T)

Fields:

- target: A symbol identifying the target variable of interest
- treatment: A NamedTuple linking each treatment variable to a value
- confounders: Confounding variables affecting both the target and the treatment
- covariates: Optional extra variables affecting the target only

Examples:

CM₁ = CM(
    target=:Y₁,
    treatment=(T₁=1,),
    confounders=[:W₁, :W₂],
    covariates=[:C₁]
)

CM₂ = CM(
    target=:Y₂,
    treatment=(T₁=1, T₂="A"),
    confounders=[:W₁],
)

TMLE.IATE — Type

IATE: Interaction Average Treatment Effect

Mathematical definition for pairwise interaction:

Eₓ[E[Y|do(T₁=1, T₂=1), X]] - Eₓ[E[Y|do(T₁=1, T₂=0), X]] - Eₓ[E[Y|do(T₁=0, T₂=1), X]] + Eₓ[E[Y|do(T₁=0, T₂=0), X]]

Assumed Causal graph:

 T  ←  W  
  ↘   ↙ 
    Y  ← C

Notation:

Y: target

T: treatment

W: confounders

C: covariates

X = (W, C, T)

Fields:

- target: A symbol identifying the target variable of interest
- treatment: A NamedTuple linking each treatment variable to case/control values
- confounders: Confounding variables affecting both the target and the treatment
- covariates: Optional extra variables affecting the target only

Examples:

IATE₁ = IATE(
    target=:Y₁,
    treatment=(T₁=(case=1, control=0), T₂=(case="A", control="B")),
    confounders=[:W₁],
)

TMLE.NuisanceSpec — Method

NuisanceSpec(Q, G; H=encoder(), F=Q_model(target_scitype(Q)))

Specification of the nuisance parameters to be learnt.

Arguments:

Q: For the estimation of E₀[Y|T=case, X]

G: For the estimation of P₀(T|W)

H: The TreatmentTransformer` to deal with categorical treatments

F: The generalized linear model used to fluctuate the initial Q

cache: Whether corresponding machines will cache data or not.

TMLE.TMLECache — Type

This structure is used as a cache for: - data: data changes - η: The nuisance parameters estimators - Ψ: The parameter of interest - η_spec: The specification of the learning algorithms used to estimate the nuisance parameters

In many cases, if some of those fields do not change, the nuisance parameters do not have to be estimated again hence saving computations.

Statistics.var — Method

var(r::ALEstimate)

Computes the estimated variance associated with the estimate.

Statistics.var — Method

var(r::ComposedTMLE)

Computes the estimated variance associated with the estimate.

TMLE.compose — Method

compose(f, estimation_results::Vararg{ALEstimate, N}) where N

Provides an estimator of f(estimation_results...).

Mathematical details

The following is a summary from Asymptotic Statistics, A. W. van der Vaart.

Consider k TMLEs computed from a dataset of size n and embodied by Tₙ = (T₁,ₙ, ..., Tₖ,ₙ). Since each of them is asymptotically normal, the multivariate CLT provides the joint distribution:

√n(Tₙ - Ψ₀) ↝ N(0, Σ),

where Σ is the covariance matrix of the TMLEs influence curves.

Let f:ℜᵏ→ℜᵐ, be a differentiable map at Ψ₀. Then, the delta method provides the limiting distribution of √n(f(Tₙ) - f(Ψ₀)). Because Tₙ is normal, the result is:

√n(f(Tₙ) - f(Ψ₀)) ↝ N(0, ∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ),

where ∇f(Ψ₀):ℜᵏ→ℜᵐ is a linear map such that by abusing notations and identifying the function with the multiplication matrix: ∇f(Ψ₀):h ↦ ∇f(Ψ₀) ̇h. And the matrix ∇f(Ψ₀) is the jacobian of f at Ψ₀.

Hence, the only thing we need to do is:

Compute the covariance matrix Σ

Compute the jacobian ∇f, which can be done using Julia's automatic differentiation facilities.

The final estimator is normal with mean f₀=f(Ψ₀) and variance σ₀=∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ

Arguments

f: An array-input differentiable map.

estimation_results: 1 or more ALEstimate structs.

Examples

Assuming res₁ and res₂ are TMLEs:

f(x, y) = [x^2 - y, y - 3x]
compose(f, res₁, res₂)

TMLE.estimate — Method

estimate(r::ALEstimate)

Retrieves the final estimate: after the TMLE step.

TMLE.estimate — Method

estimate(r::ComposedTMLE)

Retrieves the final estimate: after the TMLE step.

TMLE.optimize_ordering! — Method

optimize_ordering!(parameters::Vector{<:Parameter})

Reorders the given parameters so that most nuisance parameters fits can be reused. Given the assumed causal graph:

 T  ←  W  
  ↘   ↙ 
    Y  ← C

Notation:

Y: target

T: treatment

W: confounders

C: covariates

X = (W, C, T)

and the requirements to estimate both p(T|W) and E[Y|W, T, C].

A natural ordering of the parameters in order to save computations is given by the following variables ordering: (W, T, Y, C)

TMLE.optimize_ordering — Method

optimize_ordering(parameters::Vector{<:Parameter})

See optimize_ordering!

TMLE.parameters_from_yaml — Method

parameters_from_yaml(filepath)

Read estimation Parameters from YAML file.

TMLE.parameters_to_yaml — Method

parameters_to_yaml(filepath, parameters::Vector{<:Parameter})

Write estimation Parameters to YAML file.

TMLE.tmle! — Method

tmle!(cache::TMLECache, η_spec::NuisanceSpec, Ψ::Parameter; kwargs...)

Updates the TMLECache with η_spec and Ψ and runs TMLE.

Arguments

Ψ: The parameter of interest

ηspec: The specification for learning `Q0andG_0`

dataset: A tabular dataset respecting the Table.jl interface

verbosity: The logging level

threshold: To avoid small values of Ĝ to cause the "clever covariate" to explode

weighted_fluctuation: Fits a weighted fluctuation

TMLE.tmle! — Method

tmle!(cache::TMLECache, η_spec::NuisanceSpec; kwargs...)

Updates the TMLECache with η_spec and runs TMLE.

Arguments

Ψ: The parameter of interest

ηspec: The specification for learning `Q0andG_0`

dataset: A tabular dataset respecting the Table.jl interface

verbosity: The logging level

threshold: To avoid small values of Ĝ to cause the "clever covariate" to explode

weighted_fluctuation: Fits a weighted fluctuation

TMLE.tmle! — Method

tmle!(cache::TMLECache, Ψ::Parameter, η_spec::NuisanceSpec; kwargs...)

Updates the TMLECache with η_spec and Ψ and runs TMLE.

Arguments

Ψ: The parameter of interest

ηspec: The specification for learning `Q0andG_0`

dataset: A tabular dataset respecting the Table.jl interface

verbosity: The logging level

threshold: To avoid small values of Ĝ to cause the "clever covariate" to explode

weighted_fluctuation: Fits a weighted fluctuation

TMLE.tmle! — Method

tmle!(cache::TMLECache, Ψ::Parameter; verbosity=1, threshold=1e-8)

Updates the TMLECache with Ψ and runs TMLE.

Arguments

Ψ: The parameter of interest

ηspec: The specification for learning `Q0andG_0`

dataset: A tabular dataset respecting the Table.jl interface

verbosity: The logging level

threshold: To avoid small values of Ĝ to cause the "clever covariate" to explode

weighted_fluctuation: Fits a weighted fluctuation

TMLE.tmle! — Method

tmle!(cache::TMLECache; verbosity=1, threshold=1e-8, weighted_fluctuation=false)

Runs TMLE using the provided TMLECache.

Arguments

Ψ: The parameter of interest

ηspec: The specification for learning `Q0andG_0`

dataset: A tabular dataset respecting the Table.jl interface

verbosity: The logging level

threshold: To avoid small values of Ĝ to cause the "clever covariate" to explode

weighted_fluctuation: Fits a weighted fluctuation

TMLE.tmle — Method

tmle(Ψ::Parameter, η_spec::NuisanceSpec, dataset; kwargs...)

Build a TMLECache struct and runs TMLE.

Arguments

Ψ: The parameter of interest

ηspec: The specification for learning `Q0andG_0`

dataset: A tabular dataset respecting the Table.jl interface

verbosity: The logging level

threshold: To avoid small values of Ĝ to cause the "clever covariate" to explode

weighted_fluctuation: Fits a weighted fluctuation