API Reference

TMLE.ATE — Type

ATE: Average Treatment Effect

Mathematical definition:

Eₓ[E[Y|T=case, X]] - Eₓ[E[Y|T=control, X]]

Causal graph:

 T  ←  W  
  ↘   ↙ 
    Y  ← C

Notation:

Y: target

T: treatment

W: confounders

C: covariates

X = (W, C, T)

Fields:

- target: A symbol identifying the target variable of interest
- treatment: A NamedTuple linking each treatment variable to case/control values
- confounders: Confounding variables affecting both the target and the treatment
- covariates: Optional extra variables affecting the target only

Examples:

ATE₁ = ATE(
    target=:Y₁,
    treatment=(T₁=(case=1, control=0),),
    confounders=[:W₁, :W₂],
    covariates=[:C₁]
)

ATE₂ = ATE(
    target=:Y₂,
    treatment=(T₁=(case=1, control=0), T₂=(case="A", control="B")),
    confounders=[:W₁],
)

TMLE.CM — Type

CM: Conditional Mean

Mathematical definition:

Eₓ[E[Y|T=t, X]]

Causal graph:

 T  ←  W  
  ↘   ↙ 
    Y  ← C

Notation:

Y: target

T: treatment

W: confounders

C: covariates

X = (W, C, T)

Fields:

- target: A symbol identifying the target variable of interest
- treatment: A NamedTuple linking each treatment variable to a value
- confounders: Confounding variables affecting both the target and the treatment
- covariates: Optional extra variables affecting the target only

Examples:

CM₁ = CM(
    target=:Y₁,
    treatment=(T₁=1,),
    confounders=[:W₁, :W₂],
    covariates=[:C₁]
)

CM₂ = CM(
    target=:Y₂,
    treatment=(T₁=1, T₂="A"),
    confounders=[:W₁],
)

TMLE.IATE — Type

IATE: Interaction Average Treatment Effect

Mathematical definition for pairwise interaction:

Eₓ[E[Y|T₁=1, T₂=1, X]] - Eₓ[E[Y|T₁=1, T₂=0, X]] - Eₓ[E[Y|T₁=0, T₂=1, X]] + Eₓ[E[Y|T₁=0, T₂=0, X]]

Causal graph:

 T  ←  W  
  ↘   ↙ 
    Y  ← C

Notation:

Y: target

T: treatment

W: confounders

C: covariates

X = (W, C, T)

Fields:

- target: A symbol identifying the target variable of interest
- treatment: A NamedTuple linking each treatment variable to case/control values
- confounders: Confounding variables affecting both the target and the treatment
- covariates: Optional extra variables affecting the target only

Examples:

IATE₁ = IATE(
    target=:Y₁,
    treatment=(T₁=(case=1, control=0), T₂=(case="A", control="B")),
    confounders=[:W₁],
)

TMLE.NuisanceSpec — Method

NuisanceSpec(Q, G; H=encoder(), F=Q_model(target_scitype(Q)))

Specification of the nuisance parameters to be learnt.

Arguments:

Q: For the estimation of E₀[Y|T=case, X]

G: For the estimation of P₀(T|W)

H: The TreatmentTransformer` to deal with categorical treatments

F: The generalized linear model used to fluctuate the initial Q

TMLE.compose — Method

compose(f, estimation_results::Vararg{PointTMLE, N}) where N

Provides an estimator of f(estimation_results...).

Mathematical details

The following is a summary from Asymptotic Statistics, A. W. van der Vaart.

Consider k TMLEs computed from a dataset of size n and embodied by Tₙ = (T₁,ₙ, ..., Tₖ,ₙ). Since each of them is asymptotically normal, the multivariate CLT provides the joint distribution:

√n(Tₙ - Ψ₀) ↝ N(0, Σ),

where Σ is the covariance matrix of the TMLEs influence curves.

Let f:ℜᵏ→ℜᵐ, be a differentiable map at Ψ₀. Then, the delta method provides the limiting distribution of √n(f(Tₙ) - f(Ψ₀)). Because Tₙ is normal, the result is:

√n(f(Tₙ) - f(Ψ₀)) ↝ N(0, ∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ),

where ∇f(Ψ₀):ℜᵏ→ℜᵐ is a linear map such that by abusing notations and identifying the function with the multiplication matrix: ∇f(Ψ₀):h ↦ ∇f(Ψ₀) ̇h. And the matrix ∇f(Ψ₀) is the jacobian of f at Ψ₀.

Hence, the only thing we need to do is:

Compute the covariance matrix Σ

Compute the jacobian ∇f, which can be done using Julia's automatic differentiation facilities.

The final estimator is normal with mean f₀=f(Ψ₀) and variance σ₀=∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ

Arguments

f: An array-input differentiable map.

estimation_results: 1 or more PointTMLE structs.

Examples

Assuming res₁ and res₂ are TMLEs:

f(x, y) = [x^2 - y, y - 3x]
compose(f, res₁, res₂)

TMLE.parameters_from_yaml — Method

parameters_from_yaml(path)

Instantiate parameters described in the provided YAML file.

TMLE.tmle! — Method

tmle!(cache::TMLECache, η_spec::NuisanceSpec; verbosity=1, threshold=1e-8)

Runs the TMLE procedure for the new nuisance parameters specification η_spec while potentially reusing cached nuisance parameters.

TMLE.tmle! — Method

tmle!(cache::TMLECache, Ψ::Parameter, η_spec::NuisanceSpec; verbosity=1, threshold=1e-8)

Runs the TMLE procedure for the new parameter Ψ and the new nuisance parameters specification η_spec while potentially reusing cached nuisance parameters.

TMLE.tmle! — Method

tmle!(cache::TMLECache, Ψ::Parameter; verbosity=1, threshold=1e-8)

Runs the TMLE procedure for the new parameter Ψ while potentially reusing cached nuisance parameters.

TMLE.tmle — Method

tmle(Ψ::Parameter, η_spec::NuisanceSpec, dataset; verbosity=1, threshold=1e-8)

Main entrypoint to run the TMLE procedure.

Arguments

Ψ: The parameter of interest

ηspec: The specification for learning `Q0andG_0`

dataset: A tabular dataset respecting the Table.jl interface

verbosity: The logging level

threshold: To avoid small values of Ĝ to cause the "clever covariate" to explode