API Reference

HypothesisTests.OneSampleTTest — Type

OneSampleTTest(r::EICEstimate, Ψ₀=0)

Performs a T test on the EICEstimate.

HypothesisTests.OneSampleTTest — Type

OneSampleTTest(r::ComposedEstimate, Ψ₀=0)

Performs a T test on the ComposedEstimate.

HypothesisTests.OneSampleZTest — Type

OneSampleZTest(r::EICEstimate, Ψ₀=0)

Performs a Z test on the EICEstimate.

HypothesisTests.OneSampleZTest — Type

OneSampleZTest(r::ComposedEstimate, Ψ₀=0)

Performs a T test on the ComposedEstimate.

TMLE.Configuration — Method

Configuration(;estimands, scm=nothing, adjustment=nothing) = Configuration(estimands, scm, adjustment)

A Configuration is a set of estimands to be estimated. If the set of estimands contains causal (identifiable) estimands, these will be identified using the provided scm and adjustment method.

TMLE.OSE — Method

OSE(;models=default_models(), resampling=nothing, ps_lowerbound=1e-8, machine_cache=false)

Defines a One Step Estimator using the specified models for estimation of the nuisance parameters. The estimator is a function that can be applied to estimate estimands for a dataset.

Arguments

models: A NamedTuple{variables}(models) where the variables are the outcome variables modeled by the models.

resampling: Outer resampling strategy. Setting it to nothing (default) falls back to vanilla estimation while

any valid MLJ.ResamplingStrategy will result in CV-OSE.

ps_lowerbound: Lowerbound for the propensity score to avoid division by 0. The special value nothing will

result in a data adaptive definition as described in here.

machine_cache: Whether MLJ.machine created during estimation should cache data.

Example

using MLJLinearModels
models = (Y = LinearRegressor(), T = LogisticClassifier())
ose = OSE()
Ψ̂ₙ, cache = ose(Ψ, dataset)

TMLE.SCM — Type

A SCM is simply a wrapper around a MetaGraph over a Directed Acyclic Graph.

TMLE.TMLEE — Method

TMLEE(;models=default_models(), resampling=nothing, ps_lowerbound=1e-8, weighted=false, tol=nothing, machine_cache=false)

Defines a TMLE estimator using the specified models for estimation of the nuisance parameters. The estimator is a function that can be applied to estimate estimands for a dataset.

Arguments

models: A NamedTuple{variables}(models) where the variables are the outcome variables modeled by the models.

resampling: Outer resampling strategy. Setting it to nothing (default) falls back to vanilla TMLE while

any valid MLJ.ResamplingStrategy will result in CV-TMLE.

ps_lowerbound: Lowerbound for the propensity score to avoid division by 0. The special value nothing will

result in a data adaptive definition as described in here.

weighted: Whether the fluctuation model is a classig GLM or a weighted version. The weighted fluctuation has

been show to be more robust to positivity violation in practice.

tol: This is not used at the moment.

machine_cache: Whether MLJ.machine created during estimation should cache data.

Example

using MLJLinearModels
tmle = TMLEE()
Ψ̂ₙ, cache = tmle(Ψ, dataset)

TMLE.TreatmentTransformer — Method

TreatmentTransformer(;encoder=encoder())

Treatments in TMLE are represented by CategoricalArrays. If a treatment column has type OrderedFactor, then its integer representation is used, make sure that the levels correspond to your expectations. All other columns are one-hot encoded.

Distributions.estimate — Method

Distributions.estimate(r::ComposedEstimate)

Retrieves the final estimate: after the TMLE step.

Distributions.estimate — Method

Distributions.estimate(r::EICEstimate)

Retrieves the final estimate: after the TMLE step.

Statistics.var — Method

var(r::ComposedEstimate)

Computes the estimated variance associated with the estimate.

Statistics.var — Method

var(r::EICEstimate)

Computes the estimated variance associated with the estimate.

TMLE.StaticSCM — Method

A plate Structural Causal Model where:

For all outcomes: oᵢ = fᵢ(treatments, confounders, outcomeextracovariates)

For all treatments: tⱼ = fⱼ(confounders)

Example

StaticSCM([:Y], [:T₁, :T₂], [:W₁, :W₂, :W₃]; outcomeextracovariates=[:C])

TMLE.brute_force_ordering — Method

brute_force_ordering(estimands; η_counts = nuisance_function_counts(estimands))

Finds an optimal ordering of the estimands to minimize maximum cache size. The approach is a brute force one, all permutations are generated and evaluated, if a minimum is found fast it is immediatly returned. The theoretical complexity is in O(N!). However due to the stop fast approach and the shuffling, this is actually expected to be much smaller than that.

TMLE.compose — Method

compose(f, estimation_results::Vararg{EICEstimate, N}) where N

Provides an estimator of f(estimation_results...).

Mathematical details

The following is a summary from Asymptotic Statistics, A. W. van der Vaart.

Consider k TMLEs computed from a dataset of size n and embodied by Tₙ = (T₁,ₙ, ..., Tₖ,ₙ). Since each of them is asymptotically normal, the multivariate CLT provides the joint distribution:

√n(Tₙ - Ψ₀) ↝ N(0, Σ),

where Σ is the covariance matrix of the TMLEs influence curves.

Let f:ℜᵏ→ℜᵐ, be a differentiable map at Ψ₀. Then, the delta method provides the limiting distribution of √n(f(Tₙ) - f(Ψ₀)). Because Tₙ is normal, the result is:

√n(f(Tₙ) - f(Ψ₀)) ↝ N(0, ∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ),

where ∇f(Ψ₀):ℜᵏ→ℜᵐ is a linear map such that by abusing notations and identifying the function with the multiplication matrix: ∇f(Ψ₀):h ↦ ∇f(Ψ₀) ̇h. And the matrix ∇f(Ψ₀) is the jacobian of f at Ψ₀.

Hence, the only thing we need to do is:

Compute the covariance matrix Σ

Compute the jacobian ∇f, which can be done using Julia's automatic differentiation facilities.

The final estimator is normal with mean f₀=f(Ψ₀) and variance σ₀=∇f(Ψ₀) ̇Σ ̇(∇f(Ψ₀))ᵀ

Arguments

f: An array-input differentiable map.

estimation_results: 1 or more EICEstimate structs.

Examples

Assuming res₁ and res₂ are TMLEs:

f(x, y) = [x^2 - y, y - 3x]
compose(f, res₁, res₂)

TMLE.factorialATE — Method

factorialATE(dataset, treatments, outcome; 
    confounders=nothing, 
    outcome_extra_covariates=(),
    positivity_constraint=nothing
)

Find all unique values for each treatment variable in the dataset and generate all possible ATEs from these values.

TMLE.factorialATE — Method

factorialATE(
    treatments_levels::NamedTuple{names}, outcome; 
    confounders=nothing, 
    outcome_extra_covariates=(),
    freq_table=nothing,
    positivity_constraint=nothing
) where names

Generate a ComposedEstimand of ATEs from the treatments_levels. The components of this estimand are generated from the treatment variables contrasts. For example, consider two treatment variables T₁ and T₂ each taking three possible values (0, 1, 2). For each treatment variable, the marginal transitive contrasts are defined by (0 → 1, 1 → 2). Note that (0 → 2) or (1 → 0) need not be considered because they are linearly dependent on the other contrasts. Then, the cartesian product of treatment contrasts is taken, resulting in a 2 x 2 = 4 dimensional joint estimand:

(T₁: 0 → 1, T₂: 0 → 1)

(T₁: 0 → 1, T₂: 1 → 2)

(T₁: 1 → 2, T₂: 0 → 1)

(T₁: 1 → 2, T₂: 1 → 2)

Return

A ComposedEstimand with causal or statistical components.

Args

treatments_levels: A NamedTuple providing the unique levels each treatment variable can take.

outcome: The outcome variable.

confounders=nothing: The generated components will inherit these confounding variables.

If nothing, causal estimands are generated.

outcome_extra_covariates=(): The generated components will inherit these outcome_extra_covariates.

positivity_constraint=nothing: Only components that pass the positivity constraint are added to the ComposedEstimand

Example:

To generate a causal composed estimand with 3 components:

factorialATE((T₁ = (0, 1), T₂=(0, 1, 2)), :Y₁)

To generate a statistical composed estimand with 9 components:

factorialATE((T₁ = (0, 1, 2), T₂=(0, 1, 2)), :Y₁, confounders=[:W₁, :W₂])

TMLE.factorialIATE — Method

factorialIATE(dataset, treatments, outcome; 
    confounders=nothing, 
    outcome_extra_covariates=(),
    positivity_constraint=nothing
)

Finds treatments levels from the dataset and generates a ComposedEstimand of IATE from them (see factorialIATE(treatments_levels, outcome; confounders=nothing, outcome_extra_covariates=())).

TMLE.factorialIATE — Method

factorialIATE(
    treatments_levels::NamedTuple{names}, outcome; 
    confounders=nothing, 
    outcome_extra_covariates=(),
    freq_table=nothing,
    positivity_constraint=nothing
) where names

Generates a ComposedEstimand of IATE from treatments_levels. The components of this estimand are generated from the treatment variables contrasts. For example, consider two treatment variables T₁ and T₂ each taking three possible values (0, 1, 2). For each treatment variable, the marginal transitive contrasts are defined by (0 → 1, 1 → 2). Note that (0 → 2) or (1 → 0) need not be considered because they are linearly dependent on the other contrasts. Then, the cartesian product of treatment contrasts is taken, resulting in a 2 x 2 = 4 dimensional joint estimand:

(T₁: 0 → 1, T₂: 0 → 1)

(T₁: 0 → 1, T₂: 1 → 2)

(T₁: 1 → 2, T₂: 0 → 1)

(T₁: 1 → 2, T₂: 1 → 2)

Return

A ComposedEstimand with causal or statistical components.

Args

treatments_levels: A NamedTuple providing the unique levels each treatment variable can take.

outcome: The outcome variable.

confounders=nothing: The generated components will inherit these confounding variables.

If nothing, causal estimands are generated.

outcome_extra_covariates=(): The generated components will inherit these outcome_extra_covariates.

positivity_constraint=nothing: Only components that pass the positivity constraint are added to the ComposedEstimand

Example:

To generate a causal composed estimand with 3 components:

factorialIATE((T₁ = (0, 1), T₂=(0, 1, 2)), :Y₁)

To generate a statistical composed estimand with 9 components:

factorialIATE((T₁ = (0, 1, 2), T₂=(0, 1, 2)), :Y₁, confounders=[:W₁, :W₂])

TMLE.groups_ordering — Method

groups_ordering(estimands)

This will order estimands based on: propensity score first, outcome mean second. This heuristic should work reasonably well in practice. It could be optimized further by:

Organising the propensity score groups that share similar components to be close together.

Brute forcing the ordering of these groups to find an optimal one.